core: move new infra load files to upper directory after legacy drop
