+++ core/kt-osrd-sim-infra/src/main/kotlin/fr/sncf/osrd/sim_infra/impl/RawInfraFromRjsBuilder.kt
@@ -1,4 +1,4 @@
-package fr.sncf.osrd.sim_infra.impl.new_impl
+package fr.sncf.osrd.sim_infra.impl
 
 import fr.sncf.osrd.fast_collections.mutableIntArrayListOf
 import fr.sncf.osrd.geom.LineString
